# frontend/openapi-s3-gateway.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "S3 Gateway",
        "version": "0.1.0"
    },
    "paths": {
        "/files/": {
            "get": {
                "summary": "Get Files",
                "operationId": "get_files_files__get",
                "parameters": [
                    {
                        "name": "x-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/{file_id}": {
            "get": {
                "summary": "Get Specific File",
                "operationId": "get_specific_file_files__file_id__get",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    },
                    {
                        "name": "x-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-User-Id"
                        }
                    },
                    {
                        "name": "x-internal-source",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Internal-Source"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Specific File",
                "operationId": "delete_specific_file_files__file_id__delete",
                "parameters": [
                    {
                        "name": "file_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "File Id"
                        }
                    },
                    {
                        "name": "x-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-User-Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/upload": {
            "post": {
                "summary": "Create File",
                "description": "P\u0159ijme soubor, ulo\u017e\u00ed metadata se statusem \"uploading\" a ode\u0161le\nbin\u00e1rn\u00ed data do Haystack Node p\u0159es broker (topic: storage.write).\nVr\u00e1t\u00ed 202 Accepted \u2013 soubor JE\u0160T\u011a NEN\u00cd fyzicky ulo\u017een.",
                "operationId": "create_file_files_upload_post",
                "parameters": [
                    {
                        "name": "x-user-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-User-Id"
                        }
                    },
                    {
                        "name": "x-internal-source",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Internal-Source"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_create_file_files_upload_post"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateFile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/internal/volume/{volume_id}/objects": {
            "get": {
                "summary": "Get Volume Objects",
                "description": "Vr\u00e1t\u00ed seznam v\u0161ech nesmazan\u00fdch objekt\u016f ulo\u017een\u00fdch v dan\u00e9m svazku.\nPou\u017e\u00edv\u00e1 compact.py / Haystack Node p\u0159i kompakci.",
                "operationId": "get_volume_objects_files_internal_volume__volume_id__objects_get",
                "parameters": [
                    {
                        "name": "volume_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Volume Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/internal/bulk-update-location": {
            "post": {
                "summary": "Bulk Update Location",
                "description": "Hromadn\u011b aktualizuje haystack_offset (a p\u0159\u00edpadn\u011b volume_id) po kompakci.\nVstup: {\"updates\": [{\"object_id\": \"...\", \"volume_id\": 1, \"new_offset\": 0, \"new_size\": 1024}]}",
                "operationId": "bulk_update_location_files_internal_bulk_update_location_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Body"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_create_file_files_upload_post": {
                "properties": {
                    "bucket_id": {
                        "type": "integer",
                        "title": "Bucket Id"
                    },
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "bucket_id",
                    "file"
                ],
                "title": "Body_create_file_files_upload_post"
            },
...
            "CreateFile": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unik\u00e1tn\u00ed ID souboru"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename",
                        "description": "P\u016fvodn\u00ed n\u00e1zev souboru"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size",
                        "description": "Velikost souboru v bytech"
                    },
                    "content_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content Type",
                        "description": "MIME typ souboru"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "size"
                ],
                "title": "CreateFile"
            },
            "DeleteResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Zpr\u00e1va o v\u00fdsledku operace"
                    },
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "ID smazan\u00e9ho souboru"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "id"
                ],
                "title": "DeleteResponse"
            },
            "FileListResponse": {
                "properties": {
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/FileMetadata"
                        },
                        "type": "array",
                        "title": "Files",
                        "description": "Seznam soubor\u016f u\u017eivatele"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total",
                        "description": "Celkov\u00fd po\u010det soubor\u016f"
                    }
                },
                "type": "object",
                "required": [
                    "files",
                    "total"
                ],
                "title": "FileListResponse"
            },
            "FileMetadata": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unik\u00e1tn\u00ed ID souboru"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename",
                        "description": "P\u016fvodn\u00ed n\u00e1zev souboru"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size",
                        "description": "Velikost souboru v bytech"
                    },
                    "content_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content Type",
                        "description": "MIME typ souboru"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Created At",
                        "description": "\u010cas vytvo\u0159en\u00ed souboru"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "size",
                    "created_at"
                ],
                "title": "FileMetadata"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}